refactor: split into small page object
