+++ e2e/src/domains/category/infrastructure/pages/category-form.component.ts
@@ -1,0 +1,93 @@
+import { Page } from "@playwright/test";
+import { BasePage } from "../../../../shared/infrastructure/pages/base.page";
+
+export interface CategoryFormParams {
+  name: string;
+  icon: string;
+  type: string;
+  parent?: string;
+}
+
+export class CategoryFormComponent extends BasePage {
+  constructor(public page: Page) {
+    super(page);
+  }
+
+  async fillCategoryForm(name: string, icon: string, type: string): Promise<void> {
+    await this.page.getByTestId("input-category-name").click();
+    await this.page.getByTestId("input-category-name").fill(name);
+    this.logger.info(`Filled category name: ${name}`);
+
+    // Select icon dynamically by matching icon name or attribute
+    const iconOption = this.page.getByTestId("select-icon").locator(`div[aria-label="${icon}"]`);
+    if ((await iconOption.count()) === 0) {
+      this.logger.warn(`Icon option '${icon}' not found, selecting default icon`);
+      await this.page.getByTestId("select-icon").locator("div").first().click();
+    } else {
+      await iconOption.click();
+      this.logger.info(`Selected icon: ${icon}`);
+    }
+
+    // Select type dynamically by role option name matching the type parameter
+    const typeOption = this.page.getByRole("option", { name: type });
+    if ((await typeOption.count()) === 0) {
+      this.logger.warn(`Type option '${type}' not found, selecting default type 'Grid'`);
+      await this.page.getByRole("option", { name: "Grid" }).click();
+    } else {
+      await typeOption.click();
+      this.logger.info(`Selected type: ${type}`);
+    }
+  }
+
+  async selectParentCategory(parent: string): Promise<void> {
+    this.logger.info(`Setting parent category: ${parent}`);
+    await this.page.getByTestId("select-parent-category").click();
+
+    const parentOption = this.page.getByRole("option", { name: parent });
+    try {
+      await parentOption.waitFor({ state: "visible", timeout: 10000 });
+      this.logger.info(`Parent option '${parent}' is visible`);
+      await parentOption.click();
+      this.logger.info(`Selected parent category: ${parent}`);
+    } catch (error) {
+      this.logger.error(`Parent category option '${parent}' not found or not visible: ${error}`);
+      throw new Error(`Parent category option '${parent}' not found or not visible`);
+    }
+  }
+
+  async submitCategoryExpectError(): Promise<void> {
+    this.logger.info(`Submitting category expecting error`);
+    const [response] = await Promise.all([
+      this.page.waitForResponse(
+        (resp) =>
+          resp.url().includes("/categories") &&
+          resp.request().method() === "POST" &&
+          resp.status() >= 400 &&
+          resp.status() < 500
+      ),
+      this.page.getByTestId("button-submit").click(),
+    ]);
+    this.logger.info(`Received expected error response: status ${response.status()}`);
+  }
+
+  async submitCategoryExpectSuccess(): Promise<string> {
+    this.logger.info(`Submitting category expecting success`);
+    const [response] = await Promise.all([
+      this.page.waitForResponse(
+        (resp) =>
+          resp.url().includes("/categories") &&
+          resp.request().method() === "POST" &&
+          resp.status() === 201
+      ),
+      this.page.getByTestId("button-submit").click(),
+    ]);
+    if (!response.ok()) {
+      const status = response.status();
+      this.logger.error(`Unexpected response status: ${status}`);
+      throw new Error(`Category submission failed with status ${status}`);
+    }
+    const data = await response.json();
+    this.logger.info(`Category submitted successfully with id: ${data.id}`);
+    return data.id;
+  }
+}
